wip: refactor sidebar filters
