+++ packages/web/src/components/layouts/SearchResults/SearchResults.module.scss
@@ -35,6 +35,25 @@
   }
 }
 
+.searchResultsResource {
+  grid-column: 2 / -1;
+
+  @include mobile-and-tablet {
+    grid-column: 1;
+  }
+}
+
+.searchResultsSidebar {
+  grid-column: 1 / 2;
+  grid-row: 1 / 3;
+  align-self: start;
+
+  @include mobile-and-tablet {
+    grid-column: 1;
+    grid-row: 2;
+  }
+}
+
 .packages {
   grid-column: 2 / -1;
   display: grid;
@@ -48,80 +67,4 @@
     margin-left: -20px;
     margin-right: -20px;
   }
-}
-
-.sidebar {
-  grid-column: 1 / 2;
-  grid-row: 1 / 3;
-  display: grid;
-  grid-gap: 24px;
-  align-self: start;
-
-  @include mobile-and-tablet {
-    grid-column: 1;
-    grid-row: 2;
-  }
-}
-
-.sidebarItem {
-  &:not(:first-child) {
-    padding-top: 24px;
-    border-top: 1px solid $gray-border;
-  }
-}
-
-.sidebarItemFilter {
-  @include mobile-and-tablet {
-    display: none;
-  }
-}
-
-.sidebarItemMobileFilter {
-  display: none;
-
-  @include mobile-and-tablet {
-    display: block;
-  }
-}
-
-.mobileSelectedCounter {
-  margin-right: 6px;
-}
-
-.mobileFilterToggle {
-  display: inline-flex;
-  margin-right: 8px;
-  margin-bottom: 8px;
-
-  &:last-child {
-    margin-right: 0;
-  }
-}
-
-.mobileFiltersTop {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 12px;
-  line-height: 22px;
-}
-
-.mobileFiltersTitle {
-  display: flex;
-  align-items: center;
-  font-weight: 500;
-  color: $gray-text;
-}
-
-.mobileFiltersIcon {
-  width: 24px;
-  height: 24px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 7px;
-}
-
-.mobileFiltersReset {
-  font-weight: 500;
 }
