+++ packages/web/src/components/ui/PeopleList/PeopleList.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import styles from './PeopleList.module.scss';
+import Person from '../Person/Person';
+
+type Props = {
+  keywordsList: string[];
+  selectedKeywords: string[];
+  selectHandler: (keyword: string) => void;
+};
+
+export default function PeopleList({ keywordsList, selectedKeywords, selectHandler }: Props) {
+  return (
+    <div className={styles.authors}>
+      {keywordsList.slice(0, 4).map((person) => (
+        <Person
+          key={person}
+          name={person}
+          image='https://via.placeholder.com/36'
+          checked={selectedKeywords.includes(person)}
+          onClick={() => selectHandler(person)}
+        />
+      ))}
+    </div>
+  );
+}
