+++ packages/web/src/components/ui/SidebarCategory/SidebarCategory.module.scss
@@ -89,52 +89,3 @@
     background-color: rgba($black, 0.6);
   }
 }
-
-.authors {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.checkboxGroup {
-  display: grid;
-  grid-gap: 18px;
-}
-
-.chipSkeletonWrapper {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.chipSkeleton {
-  display: inline-flex;
-  margin-bottom: 8px;
-
-  &:not(:last-child) {
-    margin-right: 8px;
-  }
-}
-
-.personSkeleton {
-  display: inline-flex;
-  align-items: center;
-  margin-bottom: 10px;
-
-  &:not(:last-child) {
-    margin-right: 16px;
-  }
-}
-
-.personSkeletonImage {
-  flex-shrink: 0;
-  margin-right: 6px;
-}
-
-.checkboxSkeleton {
-  display: flex;
-  align-items: center;
-}
-
-.checkboxSkeletonCheck {
-  flex-shrink: 0;
-  margin-right: 14px;
-}
